+++ libsolidity/codegen/ExpressionCompiler.cpp
@@ -1068,7 +1068,15 @@ bool ExpressionCompiler::visit(MemberAccess const& _memberAccess)
 			solAssert(false, "Invalid member access to integer");
 		break;
 	case Type::Category::Function:
-		solAssert(!!_memberAccess.expression().annotation().type->memberType(member),
+		if (member == "sig")
+		{
+			FunctionType const& type = dynamic_cast<FunctionType const&>(*_memberAccess.expression().annotation().type);
+			utils().popStackElement(type);
+			/// need to store store it as bytes4
+			m_context << (type.externalIdentifier() << 224);
+		}
+		else
+			solAssert(!!_memberAccess.expression().annotation().type->memberType(member),
 				 "Invalid member access to function.");
 		break;
 	case Type::Category::Magic: